refactor: moving the test-only utilities to outside the utils package
